move setup type logic in a hook
