+++ x-pack/plugins/fleet/common/constants/setup_type.ts
@@ -6,5 +6,3 @@
  */
 
 export const AGENTLESS_POLICY_ID = 'agentless_cspm';
-
-export const AGENTLESS_FEATURE_FLAG = 'agentless';
