+++ x-pack/plugins/fleet/common/experimental_features.ts
@@ -25,6 +25,7 @@ export const allowedExperimentalValues = Object.freeze<Record<string, boolean>>(
   kafkaOutput: true,
   outputSecretsStorage: true,
   remoteESOutput: true,
+  agentless: false,
 });
 
 type ExperimentalConfigKeys = Array<keyof ExperimentalFeatures>;
