+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/create_package_policy_page/single_page_layout/hooks/index.tsx
@@ -7,3 +7,4 @@
 
 export { useDevToolsRequest } from './devtools_request';
 export { useOnSubmit } from './form';
+export { useSetupType } from './setup_type';
